+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # develop
 
+# v0.8.1 - 2012-01-09
+  * Feature: Support DTMF Input components on Asterisk
+
 # v0.8.0 - 2012-01-06
   * Feature: Expose Blather's connection timeout config when creating a Connection::XMPP
   * Bugfix: Remove some deprecated Tropo extension components